requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
httpx==0.25.1
python-dotenv==1.0.0
pillow==10.1.0
prettier==0.0.7
duckdb==0.9.2
gitpython==3.1.40
pandas==2.1.3
numpy==1.26.2
beautifulsoup4==4.12.2
pydantic==2.5.2
python-dateutil==2.8.2